+++ src/handle.cpp
@@ -35,7 +35,7 @@ std::string FieldAttributes::str() const {
     };
 
     std::string value;
-    for(unsigned short i = 0; i < static_cast<unsigned short>(arraySize(table)); i++) {
+    for(unsigned short i = 0; i < arraySize(table); i++) {
         if(hasFlag(this->value_, static_cast<unsigned short>(1u << i))) {
             if(!value.empty()) {
                 value += " | ";
