+++ src/logger.h
@@ -56,7 +56,7 @@ public:
                     EACH_LOGGING_POLICY(GEN_STR)
 #undef GEN_STR
             };
-            for(unsigned int i = 0; i < static_cast<unsigned int>(arraySize(policies)); i++) {
+            for(unsigned int i = 0; i < arraySize(policies); i++) {
                 if(getenv(policies[i])) {
                     setFlag(this->whiteList, 1u << i);
                 }
